move fetch_feed method to Client
