+++ src/client.rs
@@ -121,10 +121,19 @@ impl Client {
     }
   }
 
+  const ACCEPTED_CONTENT_TYPES: [&'static str; 6] = [
+    "application/xml",
+    "text/xml",
+    "application/rss+xml",
+    "application/atom+xml",
+    "text/html",
+    "*/*",
+  ];
+
   pub async fn fetch_feed(&self, source: &Url) -> Result<Feed> {
     let resp = self
       .get_with(source, |builder| {
-        builder.header("Accept", "text/html,application/xml")
+        builder.header("Accept", Self::ACCEPTED_CONTENT_TYPES.join(", "))
       })
       .await?
       .error_for_status()?;
@@ -134,10 +143,11 @@ impl Client {
 
     let feed = match content_type.as_deref() {
       Some("text/html") => Feed::from_html_content(&content, source)?,
-      Some("application/xml")
-      | Some("text/xml")
-      | Some("application/rss+xml") => Feed::from_rss_content(&content)?,
+      Some("application/rss+xml") => Feed::from_rss_content(&content)?,
       Some("application/atom+xml") => Feed::from_atom_content(&content)?,
+      Some("application/xml") | Some("text/xml") => {
+        Feed::from_xml_content(&content)?
+      }
       x => todo!("{:?}", x),
     };
